+++ wasm-splats/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "wasm-splats"
 version = "0.1.0"
-authors = ["Jason Sobotka <jason.sobotka@cesium.com>"]
+authors = ["Jason Sobotka <jason.sobotka@cesium.com>", "Adam Morris <adam.morris@cesium.com>"]
 edition = "2021"
 
 [lib]
@@ -11,9 +11,9 @@ crate-type = ["cdylib", "rlib"]
 default = ["console_error_panic_hook"]
 
 [dependencies]
-wasm-bindgen = "0.2.84"
-js-sys = "0.3.72"
-web-sys = { version="0.3.72", features=["console"]}
+wasm-bindgen = "0.2.99"
+js-sys = "0.3.76"
+web-sys = { version="0.3.76", features=["console"]}
 
 # The `console_error_panic_hook` crate provides better debugging of panics by
 # logging them with `console.error`. This is great for development, but requires
@@ -22,7 +22,7 @@ web-sys = { version="0.3.72", features=["console"]}
 console_error_panic_hook = { version = "0.1.7", optional = true }
 
 [dev-dependencies]
-wasm-bindgen-test = "0.3.34"
+wasm-bindgen-test = "0.3.49"
 
 [profile.release]
 # Tell `rustc` to optimize for small code size.
